+++ src/DynamoPackages/PackageManagerExtension.cs
@@ -2,16 +2,14 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Linq;
-using System.Net;
 using System.Net.Http;
 using System.Reflection;
-using System.Threading;
-using System.Threading.Tasks;
 using Dynamo.Extensions;
 using Dynamo.Graph.Workspaces;
 using Dynamo.Interfaces;
 using Dynamo.Logging;
 using Dynamo.Models;
+using Dynamo.Session;
 using Greg;
 using Greg.Responses;
 
@@ -19,25 +17,6 @@ namespace Dynamo.PackageManager
 {
     public class PackageManagerExtension : IExtension, ILogSource, IExtensionSource 
     {
-        private class NoNetworkModeHandler : DelegatingHandler
-        {
-            public NoNetworkModeHandler() : base(new HttpClientHandler())
-            {
-            }
-
-            protected override Task<HttpResponseMessage> SendAsync(
-                HttpRequestMessage request, CancellationToken cancellationToken)
-            {
-                var json = "{\"success\":false,\"message\":\"Application is in offline mode\",\"content\":null}";
-                var response = new HttpResponseMessage(HttpStatusCode.ServiceUnavailable)
-                {
-                    Content = new StringContent(json, System.Text.Encoding.UTF8, "application/json"),
-                    ReasonPhrase = "Offline Mode Enabled"
-                };
-                return Task.FromResult(response);
-            }
-        }
-
         #region Fields & Properties
 
         private Action<Assembly> RequestLoadNodeLibraryHandler;
